+++ src/next/jdbc/prepare.clj
@@ -26,12 +26,13 @@
 
 (set! *warn-on-reflection* true)
 
-(defprotocol SettableParameter :extend-via-metadata true
+(defprotocol SettableParameter
   "Protocol for setting SQL parameters in statement objects, which
   can convert from Clojure values. The default implementation just
   calls `.setObject` on the parameter value. It can be extended to
   use other methods of `PreparedStatement` to convert and set parameter
   values. Extension via metadata is supported."
+  :extend-via-metadata true
   (set-parameter [val stmt ix]
     "Convert a Clojure value into a SQL value and store it as the ix'th
     parameter in the given SQL statement object."))